@@ -1,6 +1,6 @@
 package com.rutillastoby.zoria;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.junit.Assert.*;
 
